Addition of SVN-based tests
New tests based on SVN commits added. Due to differences in detecting
files and their type, all other tests have to be changed as well to
depend the detection on the particular types of commits and the project
they are taken from. This differentiation uses the new configuration
parameters of the VariabilityChangeAnalyzer for defining the file
identification patterns.